+++ KaguyaProjectV2/KaguyaBot/Core/Commands/Currency/BuyBait.cs
@@ -79,14 +79,14 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Commands.Currency
             user.Points -= totalCost;
             user.FishBait += amount;
 
+            await DatabaseQueries.UpdateAsync(user);
+
             await SendBasicSuccessEmbedAsync($"Awesome! I've gone ahead and added `{amount:N0} bait` to your baitbox. " +
                                                         $"Happy fishing!\n\n" +
                                                         $"New total bait: `{user.FishBait:N0}`\n" +
                                                         $"New total points: `{user.Points:N0}`\n" +
                                                         $"Price per bait: `{totalCost / amount:N0}`\n" +
                                                         $"Total price: `{totalCost:N0}`");
-
-            await DatabaseQueries.UpdateAsync(user);
         }
     }
 }