+++ KaguyaProjectV2/KaguyaBot/DataStorage/DbData/Models/UserItems.cs
@@ -1,7 +1,0 @@
-﻿namespace KaguyaProjectV2.KaguyaBot.DataStorage.DbData.Models
-{
-    public class UserItems
-    {
-        public ulong UserId { get; set; }
-    }
-}